Separate inExploreView logic from Map.js into a new View component
