+++ components/View/ExploreView.js
@@ -1,0 +1,5 @@
+import Markers from "../Markers/Markers";
+
+export default function ExploreView() {
+  return <Markers />;
+}
